pyproject.toml
[tool.poetry]
name = "calculator-app"
version = "0.0.1"
description = ""
authors = ["George Reyes <grey@apache.org>"]
readme = "README.md"
packages = [{ include = "**/*.py", from = "."  }]

[tool.poetry.dependencies]
python = "^3.8"
temporalio = "^1.4.0"
python-dotenv = "^1.0.1"
boto3 = "^1.34.106"




[tool.poetry.dev-dependencies]
ruff = "^0.1.4"
pytest = "^7.1.2"
pytest-asyncio = "^0.18.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poe.tasks]
format = [{cmd = "ruff format ."}]
test = "pytest"

[tool.pytest.ini_options]
asyncio_mode = "auto"
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)"